requirements.txt
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0